in3_agent/requirements.txt
google_cloud_aiplatform==1.121.0
loguru==0.7.3
moviepy==2.2.1
opencv_python_headless==4.12.0.88
protobuf==6.33.0
pydantic==2.12.2
python-dotenv==1.1.1
Requests==2.32.5
google-cloud-aiplatform[adk,agent_engines]